@@ -1,8 +1,8 @@
-name: simple_connectivity
+name: simple_connectivity_null_safe
 description: Flutter plugin for discovering the state of the network (WiFi &
   mobile/cellular) connectivity on Android and iOS.
-homepage: https://github.com/vanethos/simple_connectivity
-version: 0.1.1
+homepage: https://github.com/alexverbitsky/simple_connectivity_null_safe
+version: 2.0.0
 
 flutter:
   plugin:
@@ -16,7 +16,7 @@ flutter:
 dependencies:
   flutter:
     sdk: flutter
-  meta: "^1.0.5"
+  meta: "^1.7.0"
 
 dev_dependencies:
   flutter_test:
@@ -26,5 +26,5 @@ dev_dependencies:
   test: any
 
 environment:
-  sdk: ">=2.0.0-dev.28.0 <3.0.0"
-  flutter: ">=1.12.13+hotfix.5 <2.0.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=2.5.0 <3.0.0"
